Move vote_if_able to helper
