+++ crates/task-impls/src/consensus/mod.rs
@@ -3,11 +3,11 @@ use std::{collections::BTreeMap, sync::Arc};
 #[cfg(not(feature = "dependency-tasks"))]
 use anyhow::Result;
 use async_broadcast::Sender;
+#[cfg(not(feature = "dependency-tasks"))]
 use async_compatibility_layer::art::async_spawn;
 use async_lock::RwLock;
 #[cfg(async_executor_impl = "async-std")]
 use async_std::task::JoinHandle;
-use committable::Committable;
 use futures::future::join_all;
 use hotshot_task::task::{Task, TaskState};
 #[cfg(not(feature = "dependency-tasks"))]
@@ -16,33 +16,32 @@ use hotshot_types::data::VidDisperseShare;
 use hotshot_types::message::Proposal;
 use hotshot_types::{
     consensus::{CommitmentAndMetadata, Consensus},
-    data::{null_block, Leaf, QuorumProposal, ViewChangeEvidence},
+    data::{QuorumProposal, ViewChangeEvidence},
     event::{Event, EventType},
-    message::GeneralConsensusMessage,
     simple_certificate::{QuorumCertificate, TimeoutCertificate, UpgradeCertificate},
-    simple_vote::{QuorumData, QuorumVote, TimeoutData, TimeoutVote},
+    simple_vote::{QuorumVote, TimeoutData, TimeoutVote},
     traits::{
-        block_contents::BlockHeader,
         election::Membership,
         network::{ConnectedNetwork, ConsensusIntentEvent},
         node_implementation::{NodeImplementation, NodeType},
         signature_key::SignatureKey,
-        storage::Storage,
-        ValidatedState,
     },
-    vid::vid_scheme,
-    vote::{Certificate, HasViewNumber},
+    vote::HasViewNumber,
 };
+#[cfg(not(feature = "dependency-tasks"))]
+use hotshot_types::{traits::storage::Storage, vid::vid_scheme, vote::Certificate};
+#[cfg(not(feature = "dependency-tasks"))]
 use jf_primitives::vid::VidScheme;
 #[cfg(async_executor_impl = "tokio")]
 use tokio::task::JoinHandle;
-use tracing::{debug, error, info, instrument, warn};
+use tracing::{debug, error, instrument, warn};
 use vbs::version::Version;
 
 #[cfg(not(feature = "dependency-tasks"))]
-use self::proposal_helpers::handle_quorum_proposal_validated;
-#[cfg(not(feature = "dependency-tasks"))]
-use crate::consensus::proposal_helpers::{handle_quorum_proposal_recv, publish_proposal_if_able};
+use crate::consensus::helpers::{
+    handle_quorum_proposal_recv, handle_quorum_proposal_validated, publish_proposal_if_able,
+    update_state_and_vote_if_able,
+};
 use crate::{
     consensus::view_change::update_view,
     events::{HotShotEvent, HotShotTaskCompleted},
@@ -53,7 +52,7 @@ use crate::{
 };
 
 /// Helper functions to handle proposal-related functionality.
-pub(crate) mod proposal_helpers;
+pub(crate) mod helpers;
 
 /// Handles view-change related functionality.
 pub(crate) mod view_change;
@@ -144,182 +143,6 @@ pub struct ConsensusTaskState<TYPES: NodeType, I: NodeImplementation<TYPES>> {
     pub storage: Arc<RwLock<I::Storage>>,
 }
 
-#[cfg(not(feature = "dependency-tasks"))]
-#[allow(clippy::too_many_arguments)]
-#[allow(clippy::too_many_lines)]
-/// Check if we are able to vote, like whether the proposal is valid,
-/// whether we have DAC and VID share, and if so, vote.
-async fn vote_if_able<TYPES: NodeType, I: NodeImplementation<TYPES>>(
-    cur_view: TYPES::Time,
-    proposal: QuorumProposal<TYPES>,
-    public_key: TYPES::SignatureKey,
-    private_key: <TYPES::SignatureKey as SignatureKey>::PrivateKey,
-    state: Arc<RwLock<Consensus<TYPES>>>,
-    storage: Arc<RwLock<I::Storage>>,
-    decided_upgrade_cert: Option<UpgradeCertificate<TYPES>>,
-    quorum_membership: Arc<TYPES::Membership>,
-    committee_membership: Arc<TYPES::Membership>,
-    instance_state: Arc<TYPES::InstanceState>,
-    event_stream: Sender<Arc<HotShotEvent<TYPES>>>,
-) -> bool {
-    use async_lock::RwLockUpgradableReadGuard;
-    use hotshot_types::utils::{View, ViewInner};
-
-    if !quorum_membership.has_stake(&public_key) {
-        debug!(
-            "We were not chosen for consensus committee on {:?}",
-            cur_view
-        );
-        return false;
-    }
-
-    let consensus = state.upgradable_read().await;
-    // Only vote if you has seen the VID share for this view
-    let Some(vid_shares) = consensus.vid_shares.get(&proposal.view_number) else {
-        debug!(
-            "We have not seen the VID share for this view {:?} yet, so we cannot vote.",
-            proposal.view_number
-        );
-        return false;
-    };
-    let Some(vid_share) = vid_shares.get(&public_key).cloned() else {
-        debug!("we have not seen our VID share yet");
-        return false;
-    };
-
-    if let Some(upgrade_cert) = &decided_upgrade_cert {
-        if upgrade_cert.in_interim(cur_view)
-            && Some(proposal.block_header.payload_commitment())
-                != null_block::commitment(quorum_membership.total_nodes())
-        {
-            info!("Refusing to vote on proposal because it does not have a null commitment, and we are between versions. Expected:\n\n{:?}\n\nActual:{:?}", null_block::commitment(quorum_membership.total_nodes()), Some(proposal.block_header.payload_commitment()));
-            return false;
-        }
-    }
-
-    // Only vote if you have the DA cert
-    // ED Need to update the view number this is stored under?
-    let Some(cert) = consensus.saved_da_certs.get(&cur_view) else {
-        return false;
-    };
-
-    let view = cert.view_number;
-    // TODO: do some of this logic without the vote token check, only do that when voting.
-    let justify_qc = proposal.justify_qc.clone();
-    let parent = consensus
-        .saved_leaves
-        .get(&justify_qc.get_data().leaf_commit)
-        .cloned();
-
-    // Justify qc's leaf commitment is not the same as the parent's leaf commitment, but it should be (in this case)
-    let Some(parent) = parent else {
-        warn!(
-            "Proposal's parent missing from storage with commitment: {:?}, proposal view {:?}",
-            justify_qc.get_data().leaf_commit,
-            proposal.view_number,
-        );
-        return false;
-    };
-    let (Some(parent_state), _) = consensus.get_state_and_delta(parent.get_view_number()) else {
-        warn!("Parent state not found! Consensus internally inconsistent");
-        return false;
-    };
-    let Ok((validated_state, state_delta)) = parent_state
-        .validate_and_apply_header(
-            instance_state.as_ref(),
-            &parent,
-            &proposal.block_header.clone(),
-            vid_share.data.common.clone(),
-        )
-        .await
-    else {
-        warn!("Block header doesn't extend the proposal!");
-        return false;
-    };
-
-    let state = Arc::new(validated_state);
-    let delta = Arc::new(state_delta);
-    let parent_commitment = parent.commit();
-
-    let proposed_leaf = Leaf::from_quorum_proposal(&proposal);
-    if proposed_leaf.get_parent_commitment() != parent_commitment {
-        return false;
-    }
-
-    if let Err(e) = storage
-        .write()
-        .await
-        .update_undecided_state(
-            consensus.saved_leaves.clone(),
-            consensus.validated_state_map.clone(),
-        )
-        .await
-    {
-        warn!("Couldn't store undecided state.  Error: {:?}", e);
-    }
-    // Validate the DAC.
-    let message = if cert.is_valid_cert(committee_membership.as_ref()) {
-        // Validate the block payload commitment for non-genesis DAC.
-        if cert.get_data().payload_commit != proposal.block_header.payload_commitment() {
-            error!(
-                "Block payload commitment does not equal da cert payload commitment. View = {}",
-                *view
-            );
-            return false;
-        }
-        if let Ok(vote) = QuorumVote::<TYPES>::create_signed_vote(
-            QuorumData {
-                leaf_commit: proposed_leaf.commit(),
-            },
-            view,
-            &public_key,
-            &private_key,
-        ) {
-            GeneralConsensusMessage::<TYPES>::Vote(vote)
-        } else {
-            error!("Unable to sign quorum vote!");
-            return false;
-        }
-    } else {
-        error!(
-            "Invalid DAC in proposal! Skipping proposal. {:?} cur view is: {:?}",
-            cert, cur_view
-        );
-        return false;
-    };
-    let mut consensus = RwLockUpgradableReadGuard::upgrade(consensus).await;
-    consensus.validated_state_map.insert(
-        view,
-        View {
-            view_inner: ViewInner::Leaf {
-                leaf: proposed_leaf.commit(),
-                state: Arc::clone(&state),
-                delta: Some(Arc::clone(&delta)),
-            },
-        },
-    );
-    if let GeneralConsensusMessage::Vote(vote) = message {
-        debug!(
-            "Sending vote to next quorum leader {:?}",
-            vote.get_view_number() + 1
-        );
-        // Add to the storage that we have received the VID disperse for a specific view
-        if let Err(e) = storage.write().await.append_vid(&vid_share).await {
-            error!(
-                "Failed to store VID Disperse Proposal with error {:?}, aborting vote",
-                e
-            );
-            return false;
-        }
-        broadcast_event(Arc::new(HotShotEvent::QuorumVoteSend(vote)), &event_stream).await;
-        return true;
-    }
-    debug!(
-        "Received VID share, but couldn't find DAC cert for view {:?}",
-        *proposal.get_view_number(),
-    );
-    false
-}
 impl<TYPES: NodeType, I: NodeImplementation<TYPES>> ConsensusTaskState<TYPES, I> {
     /// Cancel all tasks the consensus tasks has spawned before the given view
     pub async fn cancel_tasks(&mut self, view: TYPES::Time) {
@@ -331,12 +154,6 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>> ConsensusTaskState<TYPES, I>
         }
         self.spawned_tasks = keep;
         join_all(cancel).await;
-    }
-
-    /// Ignores old vote behavior and lets `QuorumVoteTask` take over.
-    #[cfg(feature = "dependency-tasks")]
-    async fn vote_if_able(&mut self, _event_stream: &Sender<Arc<HotShotEvent<TYPES>>>) -> bool {
-        false
     }
 
     /// Validates whether the VID Dispersal Proposal is correctly signed
@@ -443,18 +260,15 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>> ConsensusTaskState<TYPES, I>
                         let committee_mem = Arc::clone(&self.committee_membership);
                         let instance_state = Arc::clone(&self.instance_state);
                         let handle = async_spawn(async move {
-                            vote_if_able::<TYPES, I>(
+                            update_state_and_vote_if_able::<TYPES, I>(
                                 view,
                                 proposal,
                                 pub_key,
-                                priv_key,
                                 consensus,
                                 storage,
-                                upgrade,
                                 quorum_mem,
-                                committee_mem,
                                 instance_state,
-                                event_stream,
+                                (priv_key, upgrade, committee_mem, event_stream),
                             )
                             .await;
                         });
@@ -663,18 +477,15 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>> ConsensusTaskState<TYPES, I>
                 let committee_mem = Arc::clone(&self.committee_membership);
                 let instance_state = Arc::clone(&self.instance_state);
                 let handle = async_spawn(async move {
-                    vote_if_able::<TYPES, I>(
+                    update_state_and_vote_if_able::<TYPES, I>(
                         view,
                         proposal,
                         pub_key,
-                        priv_key,
                         consensus,
                         storage,
-                        upgrade,
                         quorum_mem,
-                        committee_mem,
                         instance_state,
-                        event_stream,
+                        (priv_key, upgrade, committee_mem, event_stream),
                     )
                     .await;
                 });
@@ -733,18 +544,15 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>> ConsensusTaskState<TYPES, I>
                 let committee_mem = Arc::clone(&self.committee_membership);
                 let instance_state = Arc::clone(&self.instance_state);
                 let handle = async_spawn(async move {
-                    vote_if_able::<TYPES, I>(
+                    update_state_and_vote_if_able::<TYPES, I>(
                         view,
                         proposal,
                         pub_key,
-                        priv_key,
                         consensus,
                         storage,
-                        upgrade,
                         quorum_mem,
-                        committee_mem,
                         instance_state,
-                        event_stream,
+                        (priv_key, upgrade, committee_mem, event_stream),
                     )
                     .await;
                 });
