+++ crates/task-impls/src/quorum_proposal.rs
@@ -32,7 +32,7 @@ use tokio::task::JoinHandle;
 use tracing::{debug, error, info, instrument, warn};
 
 #[cfg(feature = "dependency-tasks")]
-use crate::consensus::proposal_helpers::handle_quorum_proposal_validated;
+use crate::consensus::helpers::handle_quorum_proposal_validated;
 use crate::{
     events::HotShotEvent,
     helpers::{broadcast_event, cancel_task},
